+++ src/usettings.pas
@@ -6,7 +6,7 @@ interface
 
 uses
   Classes, SysUtils, Forms, Controls, Graphics, Dialogs, ButtonPanel, ExtCtrls,
-  Buttons, StdCtrls, ComboEx, utypes, uini;
+  Buttons, StdCtrls, ComboEx, utypes, uini, ugps, uaprs, mvGPSObj;
 
 type
 
@@ -15,11 +15,12 @@ type
   TFSettings = class(TForm)
     BPDefaultButtons: TButtonPanel;
     CBESymbol: TComboBoxEx;
+    cbModeSEnable: TCheckBox;
+    cbIGateEnable: TCheckBox;
     GroupBox1: TGroupBox;
     GroupBox2: TGroupBox;
     GroupBox3: TGroupBox;
     GroupBox4: TGroupBox;
-    ImageList1: TImageList;
     Label3: TLabel;
     LECleanupTime: TLabeledEdit;
     LECallsign: TLabeledEdit;
@@ -39,12 +40,16 @@ type
     SpeedButton1: TSpeedButton;
     SpeedButton2: TSpeedButton;
     SpeedButton3: TSpeedButton;
+    sbGetGPSPosition: TSpeedButton;
     procedure BBOSMMapCacheClick(Sender: TObject);
     procedure BBOSMLocalTilesClick(Sender: TObject);
     procedure BBSetDump1090(Sender: TObject);
     procedure CancelButtonClick(Sender: TObject);
+    procedure cbModeSEnableChange(Sender: TObject);
+    procedure cbIGateEnableChange(Sender: TObject);
     procedure FormShow(Sender: TObject);
     procedure OKButtonClick(Sender: TObject);
+    procedure sbGetGPSPositionClick(Sender: TObject);
   private
 
   public
@@ -56,6 +61,9 @@ var
   FConfig: PAPRSConfig;
 
 implementation
+
+Uses
+  UMain;
 
 {$R *.lfm}
 
@@ -84,6 +92,21 @@ begin
   Close;
 end;
 
+procedure TFSettings.cbModeSEnableChange(Sender: TObject);
+begin
+  LEModeSServer.Enabled := cbModeSEnable.Checked;
+  LEModeSPort.Enabled := cbModeSEnable.Checked;
+  LEModeSExecutable.Enabled := cbModeSEnable.Checked;
+end;
+
+procedure TFSettings.cbIGateEnableChange(Sender: TObject);
+begin
+  LEIgateServer.Enabled := cbIGateEnable.Checked;
+  LEIgatePassword.Enabled := cbIGateEnable.Checked;
+  LEIgatePort.Enabled := cbIGateEnable.Checked;
+  LEIgateFilter.Enabled := cbIGateEnable.Checked;
+end;
+
 procedure TFSettings.FormShow(Sender: TObject);
 var i, count: Byte;
 begin
@@ -100,6 +123,17 @@ begin
     CBESymbol.ItemsEx.AddItem(APRSAlternateSymbolTable[i].Description, i+96, 0, 0, 0, nil);
 
   CBESymbol.ItemIndex := FConfig^.AprsSymbol;
+
+  if FGPS.IsEnabled then
+    sbGetGPSPosition.Enabled := True
+  else
+    sbGetGPSPosition.Enabled := False;
+
+  cbModeSEnable.Checked := FConfig^.ModeSEnabled;
+  cbIgateEnable.Checked := FConfig^.IGateEnabled;
+
+  cbModeSEnableChange(Self);
+  cbIGateEnableChange(Self);
 end;
 
 procedure TFSettings.OKButtonClick(Sender: TObject);
@@ -112,6 +146,7 @@ begin
   FConfig^.IGatePort := StrToInt(LEIGatePort.Caption);
   FConfig^.IGatePassword := LEIGatePassword.Caption;
   FConfig^.IGateFilter := LEIGateFilter.Caption;
+  FConfig^.IGateEnabled := cbIgateEnable.Checked;
   FConfig^.CleanupTime := StrToInt(LECleanupTime.Caption);
   FConfig^.LocalTilesDirectory := LEMapLocalDirectory.Caption;
 
@@ -119,15 +154,19 @@ begin
   FConfig^.ModeSPort := StrToInt(LEModeSPort.Caption);
   FConfig^.ModeSExecutable := LEModeSExecutable.Caption;
   FConfig^.AprsSymbol := CBESymbol.ItemIndex;
-
-  if Length(FConfig^.IGatePassword) > 0 then
-    FConfig^.IGateEnabled := True;
+  FConfig^.ModeSEnabled := LEModeSServer.Enabled;
 
   SaveConfigToFile(FConfig);
-  if MessageDlg('To apply the configuration, we have to restart APRSMap.', mtConfirmation, [mbCancel, mbOk], 0) = mrOk then
-    RestartApplication;
+
+  SetPoi(FMain.PoILayer, FConfig^.Latitude, FConfig^.Longitude, FConfig^.Callsign, True, FConfig^.AprsSymbol+1, FMain.MVMap.GPSItems);
 
   Close;
+end;
+
+procedure TFSettings.sbGetGPSPositionClick(Sender: TObject);
+begin
+  LELatitude.Text := FloatToStrF(FGPS.GetLat, ffFixed, 3, 6);
+  LELongitude.Text := FloatToStrF(FGPS.GetLon, ffFixed, 3, 5);
 end;
 
 procedure TFSettings.SetConfig(Config: PAPRSConfig);